// frontend/app/model/time-slot-response.ts
/*
  - Copyright 2022 Sven Loesekann
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
    You may obtain a copy of the License at
    http://www.apache.org/licenses/LICENSE-2.0
    Unless required by applicable law or agreed to in writing, software
    distributed under the License is distributed on an "AS IS" BASIS,
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License.
*/
export interface TimeSlotResponse {
  ID:            number;
  CreatedAt:     Date;
  UpdatedAt:     Date;
  DeletedAt:     Date;
  GasStationNum: number;
  AvgE5:         number;
  AvgE10:        number;
  AvgDiesel:     number;
  GsNumE5:       number;
  GsNumE10:      number;
  GsNumDiesel:   number;
  StartDate:     Date;    
  CountyDataID:  number;
}

export interface TimeSlot {
  x: string;
  e5: number;
  e10: number;
  diesel: number;
}